fix: check that `onSuccess()` is called only once

+++ app/src/main/java/com/github/onlynotesswent/model/folder/FolderRepositoryFirestore.kt
@@ -55,7 +55,11 @@ class FolderRepositoryFirestore(
       withContext(Dispatchers.IO) { folderDao.addFolder(folder) }
     }
 
-    Log.e(TAG, "Adding folder: $folder")
+    if (!NetworkUtils.isInternetAvailable(context)) {
+      onSuccess()
+      return
+    }
+
     db.collection(folderCollectionPath).document(folder.id).set(folder).addOnCompleteListener {
         result ->
       if (result.isSuccessful) {
@@ -78,6 +82,11 @@ class FolderRepositoryFirestore(
     // Update the cache if needed
     if (useCache) {
       withContext(Dispatchers.IO) { folderDao.addFolders(folders) }
+    }
+
+    if (!NetworkUtils.isInternetAvailable(context)) {
+      onSuccess()
+      return
     }
 
     val batch = db.batch()
@@ -108,6 +117,11 @@ class FolderRepositoryFirestore(
       withContext(Dispatchers.IO) { folderDao.deleteFolderById(folderId) }
     }
 
+    if (!NetworkUtils.isInternetAvailable(context)) {
+      onSuccess()
+      return
+    }
+
     db.collection(folderCollectionPath).document(folderId).delete().addOnCompleteListener { result
       ->
       if (result.isSuccessful) {
@@ -130,6 +144,11 @@ class FolderRepositoryFirestore(
     // Update the cache if needed
     if (useCache) {
       withContext(Dispatchers.IO) { folderDao.deleteFoldersFromUid() }
+    }
+
+    if (!NetworkUtils.isInternetAvailable(context)) {
+      onSuccess()
+      return
     }
 
     db.collection(folderCollectionPath).get().addOnCompleteListener { task ->
@@ -279,7 +298,11 @@ class FolderRepositoryFirestore(
     // Update the cache if needed
     if (useCache) {
       withContext(Dispatchers.IO) { folderDao.addFolder(folder) }
+    }
+
+    if (!NetworkUtils.isInternetAvailable(context)) {
       onSuccess()
+      return
     }
 
     db.collection(folderCollectionPath).document(folder.id).set(folder).addOnCompleteListener {
