+++ app/src/main/java/com/github/onlynotesswent/model/note/NoteRepositoryFirestore.kt
@@ -199,6 +199,11 @@ class NoteRepositoryFirestore(
       withContext(Dispatchers.IO) { noteDao.addNote(note) }
     }
 
+    if (!NetworkUtils.isInternetAvailable(context)) {
+      onSuccess()
+      return
+    }
+
     performFirestoreOperation(
         db.collection(collectionPath).document(note.id).set(convertNotes(note)),
         onSuccess,
@@ -214,6 +219,11 @@ class NoteRepositoryFirestore(
     // Update the cache if needed
     if (useCache) {
       withContext(Dispatchers.IO) { noteDao.addNotes(notes) }
+    }
+
+    if (!NetworkUtils.isInternetAvailable(context)) {
+      onSuccess()
+      return
     }
 
     val batch = db.batch()
@@ -233,7 +243,11 @@ class NoteRepositoryFirestore(
     // Update the cache if needed
     if (useCache) {
       withContext(Dispatchers.IO) { noteDao.addNote(note) }
+    }
+
+    if (!NetworkUtils.isInternetAvailable(context)) {
       onSuccess()
+      return
     }
 
     performFirestoreOperation(
@@ -253,6 +267,11 @@ class NoteRepositoryFirestore(
       withContext(Dispatchers.IO) { noteDao.deleteNoteById(id) }
     }
 
+    if (!NetworkUtils.isInternetAvailable(context)) {
+      onSuccess()
+      return
+    }
+
     performFirestoreOperation(
         db.collection(collectionPath).document(id).delete(), onSuccess, onFailure)
   }
@@ -266,6 +285,11 @@ class NoteRepositoryFirestore(
     // Update the cache if needed
     if (useCache) {
       withContext(Dispatchers.IO) { noteDao.deleteNotesFromUid() }
+    }
+
+    if (!NetworkUtils.isInternetAvailable(context)) {
+      onSuccess()
+      return
     }
 
     db.collection(collectionPath).get().addOnCompleteListener { task ->
@@ -333,6 +357,11 @@ class NoteRepositoryFirestore(
     // Update the cache if needed
     if (useCache) {
       withContext(Dispatchers.IO) { noteDao.deleteNotesFromFolder(folderId) }
+    }
+
+    if (!NetworkUtils.isInternetAvailable(context)) {
+      onSuccess()
+      return
     }
 
     db.collection(collectionPath).get().addOnCompleteListener { task ->
